Refactored IFont and text geometry generation
Closed #467

+++ src/gamebase/include/gamebase/impl/text/AlignedString.h
@@ -11,18 +11,18 @@
 namespace gamebase { namespace impl {
 
 struct AlignedString {
-    AlignedString(const BoundingBox& bbox, std::vector<size_t>&& glyphIndices)
+    AlignedString(const BoundingBox& bbox, std::vector<uint32_t>&& glyphIndices)
         : bbox(bbox), glyphIndices(std::move(glyphIndices))
     {}
 
-    AlignedString(const BoundingBox& bbox, const std::vector<size_t>& glyphIndices)
+    AlignedString(const BoundingBox& bbox, const std::vector<uint32_t>& glyphIndices)
         : bbox(bbox), glyphIndices(glyphIndices)
     {}
 
     const Vec2& offset() const { return bbox.bottomLeft; }
 
     BoundingBox bbox;
-    std::vector<size_t> glyphIndices;
+    std::vector<uint32_t> glyphIndices;
 };
 
 } }
