+++ src/gamebase/src/impl/text/FontMetaData.h
@@ -19,16 +19,16 @@ public:
 
     FontMetaData(int firstGlyph, int lastGlyph);
 
-    std::vector<size_t> glyphIndices(const std::string& utf8Str) const;
+    std::vector<uint32_t> glyphIndices(const std::string& utf8Str) const;
         
     virtual void serialize(Serializer& s) const override;
 
 private:
-    size_t getGlyphIndex(uint64_t utf8CharCode) const;
+    uint32_t getGlyphIndex(uint64_t utf8CharCode) const;
 
     friend std::unique_ptr<IObject> deserializeFontMetaData(Deserializer&);
 
-    std::unordered_map<uint64_t, size_t> m_glyphIndices;
+    std::unordered_map<uint64_t, uint32_t> m_glyphIndices;
 };
 
 } }
